[FIX] sap.m.ViewSettingsDialog: Cloned item now works with binding
When a cancel button is clicked we restore the previous state of the
ViewSettingsDialog through cloned item. However, when a clone is used
the binding is cloned but only the context without the values.
In order to make values appear it is necessary to add the control in
the model propagation tree (this can be done by adding the cloned item
to the dependents aggregation).

Change-Id: I9d6b577fc7a599e953afe41f1a32d218d368683f
BCP: 1770487283
CR-Id: 002075125900003822442017