[INTERNAL] DemokitV2: References links improvement
The problem:
Where there was only one reference link coming from the JSDoc and
it was extracted in the header on it's place there was a link
with text undefined rendered.

The solution:
Traversing the reference links now is with a for loop adding to the
array instead of using array.pototype.map which can add undefined as
a entry to the final array.

Additional:
Title for reference links renamed from "References" to
"Documentation links"

Change-Id: I1a2bd10e9385ccbef7606299be347527f21bed50