[INTERNAL] sap.f.DynamicPage, sap.uxap.ObjectPageLayout transparency fixed
Issues addressed:
1. DynamicPage header - whole content fades (not just the background)
Background: when in the "sap_belize_plus" theme in FLP (sapUshellShell)
the sap.f.DynamicPage header area becomes translucent.

The problem: the translucency was achieved by "opacity: 0.6" which
also makes the text and images inside the whole header area fade.

The solution: the "opacity" style is replaced by "background"
with "0.6" for the alpha channel, as was done for ObjectPage already.

In addition, the ObjectPage title and header, based on the DynamicPage
ones, were aligned with the change.

2. There should be no bottom border for sap.f.DynamicPageHeader in the
object page use case. This was already fixed for belize, now also fixed
for belize_plus ("!important" was used to avoid overly complex selectors).

3. The fade percentage for ObjectPage was changed from 40% to 60%
because the fade less function takes as an argument an absolute value
(and not by how many % to decrease the opacity).

4. The linear gradient, used for anchor bar left/right arrows removed
for the transparency scenario as it relies on a solid color, which is
absent in this use case.

Change-Id: I6559a7ca5dd969bf2ec1648e2f0df79479c6e0d6