[FIX] sap.ui.layout.Grid: Margin for hidden elements removed
When a given block is hidden, it still takes up some space.
This is because the internal grid's elements still have some
margins, despite the element's visibility. This fix adds
a class, which overrides the default margin in onAfterRendering,
and then, with the help of an observer, it watches for changes
on those elements. If one element's visibility is changed - it will be
the only one to be updated.

Change-Id: Ib4e7bb03259010ea3e3b467b0e8b502d4c4f2669
BCP: 1770443884