[INTERNAL] Property bag for initial file content to set generator (tool)
If generator is provided in property bag in
Change.createInitialFileContent/Variant.createInitialFileContent and
ContextManager createOrUpdateContextObject, the attribute
support/generator is filled with that value.

If sTool is provided in DescriptorChangeFactory.createNew, the property
bag property generator is filled when calling
Change.createInitialFileContent. With that descriptor changes contain
the tool under attribute support/generator. 

Change-Id: I8441333de1041a4c5b90845bad4493bd05ed34c8