[FIX] sap.ui.unified.FileUploader: XHR Upload on Safari restored
The problem:
On Safari browser upload of standard files using XHR may fail on
certain file types.

The root cause:
Calling FormData.append method using the third argument for file name
when passed file is not a Blob which has a name property causes
Safari browser to throw an exception and the upload fails.

The solution:
Calling FormData.append method only when the passed file is instance
of Blob and has a 'name' property

BCP: 1770523801

Change-Id: I5fc64d205f2a6b122f26561f7046fdd0b7664582