[INTERNAL][FIX] sap.m.NavContainer: page remains hidden upon navigation corrected
*Background*
The removal of CSS classes upon transition end of the child pages changed
after the following change: #2822324
which has been done to fix an edge case, when the QuickView control
(internally NavContainer is used) re-opens during navigation.
Basically the removal of classes became conditional: if (oPage.isActive())

*Issue*
However, it turns out that this change leads to another issue.
When the app uses different type of transition to navigate between two pages,
the first (from page) page remains hidden.
For example:
navigate.to(Page2) with FLIP
navigate.to(Page1) with SLIDE

*Solution*
Condition now becomes: if (oPage.getParent().isActive()).
This still work for the edge case, when the QuickView control re-opens
during navigation and will work for the current one as it becomes less restrictive.

Change-Id: I35145acd981ba6b6c472cc2122a996f9854728ad
BCP: 1770522232
CR-Id: 002075125900004079472017