[INTERNAL][FIX] sap.ui.documentation: Ensured entities have correct library name
The entities data did not contain a field for the library name.
So in the cases where the library name was needed, it was inferred from the entity namespace.

Problem:
The library name of an entity does not always coinside with the entity namespace => incorrect library name was inferred.

Fixed by ensuring the actual library name is accessed and used where needed.

Ensured the correct library name is accessed and  used.

Change-Id: I93a5837efaec68138776c84908372dac1b52ba43
BCP: 1770472795
(cherry picked from commit 44a3273f38282ba0e9d4182ecc9a743d48c504a5)
CR-Id: 002075125900004240352017