[FIX] sap.ui.layoyut.Grid: Added min-height for wrappers
The wrapper for each control inside the Grid is using
float: left for arrangement. If, however, the height of that
wrapper becomes 0, the next element will take it's place
(and this is standard html/css behavior). This is why we add
min-height to each wrapper, so that it will always take up its
space, and not lose it under any unpredictable circumstances.

See the link below for reference:
https://codepen.io/anon/pen/eeOMJg

Change-Id: I25a39ba45774ce3c0647a6293ac52ac95cfbaa8f
BCP: 1770523047