[FEATURE] core.Component: async loading of component resource bundles
Yet, the async loading of app resource bundles has been part of the
manifest model preload within the sap.ui.component function. With this
change the async loading of the ResourceModels is not part of the preload
mode anymore and happens after the preloading of all the other model types
has finshed.

The async loading of the app resource bundles is active only if the
sap.ui.component is set to "async=true" and the "manifest" property
is given.

Change-Id: I2ca6ce2774e3e44d9b776c539a31b1570b162b9a