[INTERNAL] sap.ui.documentation: master tree updated correctly
Issue
Using some of the links in the DEMOKIT`s Documentation details view,
opens the correct topic, but the master tree is not being updated.

Root cause
The issue can be reproduced with links, which have href attribute, ending with '.html'.
The new topic is opened and the router fires _onTopicMatched.
The _onTopicMatched method provides the topic ID to the master tree to allow the tree to update its selection.
Then, the master tree tries to match every node`s key with the topic ID.
However, the nodes key (in the tree model, based on the index.json) includes the part without the extension - '.html'.

Solution
preprocess the topic ID, before updating the tree selection, cutting the unneeded extension.

Another solution might be to process the href attributes of those links at build time.
Feel free to share your thoughts.

BCP: 1770537342
Change-Id: I46ecdfdad01a16019072058e7157d8f899d7c53a