[INTERNAL][FIX] sap.f.DynamicPage: determining scroll position corrected
*Issue
On zoom (110% or higher) the DynamicPage doesn not snap correclty on clicking the Title,
which triggers scrolling to the snapping point.
This can be reproduced both with Chrome and IE11 at least,
using the DynamicPageFreeStyle Explored sample.

*Background
To snap or not, the DynamicPage decides, based on the followoing condition:
this._getScrollPosition() >= this._getSnappingHeight()

This basically means,
if the scroll position is equal or bigger than the snapping point - snap.

When a click on title is performed, the snapping point is calculated,
it`s basically the position where the page should scroll to completeley scroll out the Header.
Hence, it`s equal to the Header`s height.

Example:
if the Header`s height is 100 -> snapping point is 100 (The page should scroll to this pos)
if the Header`s height is 100.3-> snapping point is 101 (The page should scroll to this pos)

*Root cause
The scrolling is delegated to the ScrollEnablement, which internally uses
this._$Container.scrollTop(y) to scroll the container.
We have experienced unexpected results when setting and getting values via scrollTop on zoom.

For example (Chrome, zoom: 110%):
set: this._$Container.scrollTop(111),
get: this._$Container.scrollTop() is 110.90908..
or
set: this._$Container.scrollTop(110.9)
get: this._$Container.scrollTop() is 110

Althogh we are setting the correct scroll position,
the getter of scrollTop returns different values on different zoom with a margin less than 1 px.
Although a small difference, it leads to wrong output of the above mentioned condition
and as a result, there is no snapping.

This is something that can`t be influenced on the lowest level
and should be handled by the DynamicPage.

*Solution
- Header`s height is now determined by getBoundingClientRect().
- The snapping point becomes Math.ceil(headerDom.getBoundingClientRect().height)
- The current scroll position is determined by Math ceil($wrapper.scrollTop())
- unit tests are adjusted

These changes produces the following result:

- if we set the scroll position to 111 and expect 111,
  and although we might get 110.9 on different zoom - after ceiling the result is 111;

- Instead of setting 110.9, we will set 111 and expect 111
  and if we get something less than 111 - after ceiling the result is 111;

BCP: 1770375061

Change-Id: Icd9c65589fabeea50811fa2fb043eae5fdf6b98f