[INTERNAL][FIX] sap.f.DynamicPage - removed redundant css
Background:
This change addresses a problem with IE11 when
reading css rules with Javascript using the "cssText" DOM API.
Calling "cssText" on a "@keyframes" rule with vendor-specific
prefixes inside leads to a "Member not found" error.
Since there already are vendor-prefixed keyframe rules, having
vendor prefixes inside the default keyframe rule is redundant.

Note: This is a downport

Change-Id: If82dffa7fe0799b6e630086e4b970d50349ae4ec
BCP: 0020283765 0000489097 2017
CR-Id: 002075125900004248832017