[INTERNAL] DemokitV2: Pre-process topic links in deprecation messages
The problem:
Link to topics in deprecation messages were not processed correctly
and were treated as a method link. All links in deprecation messages
were opening in a new window.

The root cause:
* Links were pre-processed with different method than used for overview
and description text which was not handling all possible scenarios.
* Deprecation messages are viewed using sap.m.MessageStrip control which
internally uses sap.m.FormattedText. Passing a link to FormattedText
without a target attribute it is set automatically to "_blank" by
the control.

The solution:
* All links to deprecation messages are pre-processed with the method
that handles overview and description messages.
* Method that handles link pre-processing now creates links with target
set to "_self" which is the default target attribute value.

Change-Id: I5469d40dd792b43b6bb371ffd3812935d16a3ad5