[FIX] sap.m.Menu: after removal from items aggregation item is still usable.
sap.m.Menu works with 2 sets of items - sap.m.MenuItem(public)
and sap.ui.unified.MenuItem(internal). The latter is used for visualising
menu to the end-user. The connection between both is maintained
internally.

Removing of an item from the public aggregation <items> used to destroy
the internally connected one. According to the ManagedObject's API
the removeAggregation does not destroy the given aggregation elements,
but just removes them as child.

After removing an element from an aggregation, this element should be
still usable.

Change-Id: Ib1df791707b7da1b2a6a732d28d7a68ab97a7140
BCP: 1770492569