[FIX]sap.ui.rta: Missing binding context causes exception starting RTA
When starting RTA immediately after doing a personalization change, and
only under certain circumstances, the binding context for some movable
elements is not available, causing an exception. This fix ensures that
the binding context is only checked if available. Since this check
happens several times, we can be confident that the functionality won't
be hindered as the binding will be eventually loaded.

Change-Id: Ic4558cddc7f88a7fec82c2211655ad84f2bf1c34
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>