[INTERNAL] odata.lib._V2MetadataConverter: refactoring
Create a special V2 annotatable class which maintains the unprocessed
SAP attributes and which is able to set annotations for the
correct target (automatically creating it if necessary).

This avoids that all conversions must go through convertAnnotations
which creates the target in advance and in the end has to check whether
annotations have been defined and copy it to convertedV2Annotations in
this case.

PS1: introduce and use Annotatable class, but do not restructure yet
PS2: rebased
PS3: review comments to PS1; some reordering and renaming
PS4: convertAnnotations abandoned
PS5: review comments
PS6: review comments; iterate over the v2 annotations in v2annotatable
PS7-9: review comments

Change-Id: Ic29524faba965509a73d2e740d6465e3df8569fa