[INTERNAL][FIX] uxap.ObjectPagesubSection: onbeforeRendering aborted
Issue
Coding in _applyLayout method fails as it tries to use
the parent ObjectPage instance, which is null

Root cause
The SubSection, which is invalidated, is part of a Section, that has been removed.
Invalidation process continues and leads to that error.

Solution
Stop the onBeforeRendering (already done in onAfterRendering),
when parent ObjectPage is not present.

Change-Id: I2cde05a37f697e88ba74304c9ae193b52e277df6
BCP: 1770477248