[FIX] sap.ui.unified.CalendarUtils: getFirstDateOfWeek when end/begin of year.
The US weeks at the end of December and at the beginning of January(53th and 1st),
are not considered.

If a given date is in the beginning of January (e.g. Friday, 2 Jan 2015,
 week 1),according to US rules, those are actually 2 weeks:
 - week 53(Sunday, 28 Dec 2014 - Wednesday, 31 Dec 2014) and
 - week 1(Thursday, 1 Jan 2015 - Saturday, 3 Jan 2015)

The rules above are not considered. If one wants to get it like it, she/he
can use UniversalDate.getFirstDateOfWeek

The way this function works is the following:
If a given date is in the beginning of January (e.g. Friday, 2 Jan 2015,
week 1), the function will return week start date in the previous year
(e.g. Sunday, 28 Dec 2014, week 53).

(cherry picked from commit 3f79920)

BCP: 1770534428
Change-Id: I2841eeb0fbdb5d852085233bd595842b17f009b0
CR-Id: 002075125900004425452017