[INTERNAL][FIX] sap.ui.documentation: Improve search in Samples section
The problem:
When you search for a control typing more than one word without white
space for a certain control name compound from more than one word the
search fails even if there a control with the same name exist

The root cause:
There was an improvement for the old Explored app introduced with Gerrit
change: 2399456 which fixes incorrect behavior processing search tags
which removed only the first white space but left possible other
un-removed. The improvement was not transfered in the new DemoKit
Samples section.

The solution:
Adapt the solution from the old Explored app to the new Demokit Samples
section which in general does is replace all empty string occurrences in
the search tags generated.

Change-Id: I0b36e753dfe0a4593a757873509465aea6fbbcdb