[INTERNAL][FIX] sap.m.SelectList: Multi-column list focused rows misalignment corrected
BCP: 1770583325
Did some testing with different approaches for IE, but I couldn't find a
way of retaining the focus on left/right for the Multi column select.
The reason is when display: table-row is used, the IE hack doesn't work,
so I applied the exclusion to focused rows also as it is for the
selected one.

Change-Id: I6d7ad10d1766f7878d1d614ed1b7f607849d8b30