[FIX] TimePicker-closing before slider scrolling is complete works.
If we have a sliders opened and scroll it several times in a row(via
mouse or finger) and close the dialog (click OK) before the slider
animation is completed, the TimePicker used to go into an endless loop.

The reason for this was TimePickerSlider calculation which relies on a
dom element's height to decide which item(hour,minute,second) to select.
Since the slider is closed, height cannot be obtained and the private
method _scrollerSnapped is called with parameter 'iCurrentItem' = Infinity

Change-Id: I8b18600b6d943925cecafcc32c400629fcd91683
BCP: 002028376500005440102017
CR-Id: 002075125900004434582017