[FIX] Form* tests: fix race condition in integration tests
All the Form* integration tests load their mock data asynchronously in
the onInit handler, but tried to access the data without further checks
in their 'handleEditPress' handler. So when the data had not arrived
when the button was pressed, a runtime error occurred.

This change sets the button initially to enabled=false and enables it
only after the data has been loaded. The tests have been adapted to wait
for enabled=true before pressing the button.

Change-Id: Idf00e91fa2b5272e5db37d24ee208ba56e049ad5