[INTERNAL][FIX] sap.ui.support.TechnicalInfo: Support Assistant Default location is now fixed.
BCP: n/a

The problem: When TechInfo dialog was loaded from version below 1.48 the
default location was still set to search the Support Assistant in current applicaton version
which was a problem because the Support Assistant is not available in versions below 1.48.

Solution: An additional check was added that sets default location to OpenUI5 (Nightly)
if UI version is below 1.48.

Change-Id: I678e66b2b9de47e7c38b5fc1aa31bbf5ecf1c970
CR-Id: 002075125800004447802017