[FIX] sap.m.ViewSettingsDialog: when model is changed, filter counter is updated.
When the model connected to the filter details is updated the VSD used
not to update its counters for the selected filter sub-items.

Now this is fixed, as update in the model would mean change in the
selected items.

(cherry picked from commit b417b14)

1 test module is merged:
'Misc' is merged into 'Others'
(cherry picked from commit e4dbc1e)

Change-Id: I844ce2ad5829fe64ba0651706e42f84a7c3be2ea
BCP: 002075129400005236012017
CR-Id: 002075125900004471522017