[FIX] sap.m.ViewSettingsDialog: when model is changed, filter counter is updated.
When the model connected to the filter details is updated the VSD used
not to update its counters for the selected filter sub-items.

Now this is fixed, as update in the model would mean change in the
selected items.

2 test modules are merged:
'Misc' is merged into 'Others'
'Screen Reader' is merged into 'Accessibility'

Change-Id: I844ce2ad5829fe64ba0651706e42f84a7c3be2ea
BCP: 002075129400005236012017
CR-Id: 002075125900004471552017