[FIX]sap.ui.rta: Exiting RTA removes application content from Popup
RTA has some controls rendered in the static UIArea. Calling destroy on
these elements when exiting RTA triggers the invalidation of the static
UIArea, which potentially removes application content from the screen as
well (e.g. in a Popup). The solution is to call the destroy with
suppress invalidate = true for these elements.

Change-Id: I5172172bb64da857a6ca8614934ee56caf8be6d9
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>