[FIX] Fallback for libraries without custom theme
In case a theme for a library can not be loaded, a fallback theme will
be loaded (typically the theme that the current one inherits from).
The fallback will only be done for the affected libraries, not all libs
in general.
This only applies for custom themes that are generated with certain
metadata containing that "base" theme information.

CR-Id: 002075125800004714572017
BCP: 0020751295 0000382576 2017
Change-Id: Ib1d9105091c645d86ce6d80694b8b019fd8fb4bf
(cherry picked from commit 0bc9ac6cbb8082f7181654833c1885c72319b0a7)