[FIX] sap.ui.unified.Calendar: select date on mouseup when used on touch device
On touch devices when someone touches the screen it was selecting a date,
because the mousedown event used to select date. However, it makes more
sense to select a date when mouseup happens, so on touch devices
date selections happens now on mouseup.

On desktop devices which do not support touch we keep the current behaviour,
because of singleIntervalSelection property.

Fixes: https://github.com/SAP/openui5/issues/1750

Change-Id: Idb23d8afc41b0cbed982148a37e8440909bc93b7
BCP: 002028376500005552752017
CR-Id: 002075125900004559962017