[FIX] Component/Manifest: Remove css resources on destroy
When destroying a component, all included css files should be removed
again. This behavior relied on a matching "href" of the <link> tags.
In some scenarios the URLs are adopted to include a parameter to
force cache invalidation. In that case the href didn't match anymore
with the one defined in the manifest which results into not removing the
css and including it again when an component instance is created again.

As all <link> tags are marked with the manifest's unique ID, the removal
can be done a lot easier by removing all elements regardless of their
"href" attribute.

Change-Id: I840ef4edbde4c69c1854f43754cfdc28a1039b05
BCP: 1770602032