[INTERNAL] Demokit V2: Switching both Theme and Content Density fix
The problem:
In the sap.ui.documentation project on the "Samples" section we have a
configuration where you can change the Theme and also the content
density of the application. Switching both of them on one pass sometimes
causes the browser to stall and become completely unresponsive.

The root cause:
* We have a listener attached to "onThemeChanged" event which scrolls
the tree to a certain position on theme changed.
* When you change both theme and content density the code calls
"applyTheme" to change the theme and "notifyContentDensityChanged" in
direct succession to notify controls which need to react on the content
density change.
* As the application is large there is a race condition in the core where
"notifyContentDensityChanged" calls "fireThemeChanged" when the theme
is still in transition and the CSS is not loaded which causes on our
side "ScrollEnablement" to be called with no theme and that's what is
causing the crash.

The solution:
* Make theme/density/rtl mode config changes only if we have a real change
in settings.
* Call "notifyContentDensityChanged" method only if there is no preceding
call to "applyTheme" method removing the possibility for "onThemeChanged"
event to be called in a state where no theme is currently loaded.

Change-Id: I5126de848d086e81b6bb3ac126d3ae8a9975d122