[FIX] sap.m.FacetFilter: Scroll of the items work with arrows
Previously if we have more FacetFilterLists in the FacetFilter and
try to scroll with the arrow buttons it not awlays works.

The reason for this is if the first list is focused and click
'next' arrow it will scroll the FacetFilter and hide the first
FacetFilterList. By clicking the 'next' arrow again the browser
will restore the focus to the first FacetFilterList which will
reset the scroll position of the container and handler for 'next'
arrow will do nothing but add again to the same number.

Change-Id: Iaa4cd5b9ad56fa0f3b2f7313e0157b7f7b1c484a
BCP: 1770531260
CR-Id: 002075125800004227132017