[FEATURE] TreeTable: API to expand/collapse multiple rows added
The collapse/expand methods have been enhanced to accept an array of
indices of the rows which will be expanded or collapsed. This is done
in an optimized way, where a change event will only be fired for the
last operation. Therefore, the table updates its content only after all
operations have completed.

Change-Id: I2e04fcda9c0602e1b3f713629ba5b0f9a3632a07
Resolves:
JIRA: CPOUIFTEAMB-835