[FIX] Table: Correctly calculate the number of displayed rows
The available space, which is used to set the height of the rows
container element and calculate the number of rows to render, is now
calculated in steps of 5. This should avoid issues where the available
height differs only by a few pixels (due to many reasons), resulting in
rendering one row more or less during the still active rendering
process.

Change-Id: I8acaa4da3742ddd457b57391ac98559bea9f009b
BCP: 1770615846
CR-Id: 002075125800004267102017