[FIX] core/XMLTemplateProcessor: pass correct container to ExtensionPoint
If an ExtensionPoint is declared inside of an html node in an XMLView,
the extension could not be found due to the missing view name on the
generated XMLView representing the html node. This fix takes care of
passing the containig view to the ExtensionPoint.

Change-Id: I846e8b62e3ba26883c048252b0a53c17b9234019
BCP: 0020751294 0000540578 2017
CR-Id: 002075125800004481192017
(cherry picked from commit 68c17d282bbff2a83e367fd6b7fb0d8131ae625a)