[FIX] sap.m.DatePicker - setMin/MaxDate does not change other property.
When min or max date is set to a value that makes the current dateValue
or value invalid (does not match the new min/max range), the DatePicker
used to implicitly change the property dateValue, respectively value.

This is not allowed anymore. Whatever the app. developer gives to any or
the properties minDate, maxDate, dateValue or value, it will be set as is
(unless is of wrong typ like non JavaScript date object for minDate).

With this change, value and dateValue are always in sync (even if they
don't match the min/max range). Previously in case of validation issue
during setting of a value, the dateValue used not to be synchronized,
but left with its old setting.

Since the order of calling the setters is not reliable when the properties
are data bound, the only feasible validation between the min/max date and
the current dateValue happens at before rendering phase. If validation fails,
an error is logged into the browser console, informing him/her that he/she
did something wrong and need to fix it.

Change-Id: I1edc9006ae018bd21b18bc38f8dd3ec8f5e188f3
BCP: 1770608708