[FIX] Filter/Sorter: Get comparator methods right
Comparator methods cannot be the same for sorting and filtering. For
sorting null/undefined values are ordered on the top in ascending order,
so the comparisons with null must treat it as less than any other
value. For filtering null/undefined values must be treated in a way,
that they are neither less or greater than other values.

Updated default comparator methods in Sorter/Filter to reflect this
behaviour correctly. Added a wrapper around the OData comparator methods
in v2.ODataListBinding for clientside sorting. 

BCP: 0020751295 0000379313 2017
Change-Id: I6532a3f23be968ab396ae2a94ad18074c17fa891
CR-Id: 002075125800004597152017