[INTERNAL][FIX] sap.m.Select: Ensured fire latest change
The change refactors the sap.m.Select control to prevent the problem exemplified in the specified qunit:

Problem:
When the user changes the selection with tab/click,
a select event was fired BEFORE the internal transition to the newly selected state completed.

Solution overview:
Fire a select event only AFTER the internal transition to the newly selected state completed.

Solution details:
-- Reverted change I9f99c9831112086daed7da2f126d622f43939184 [that introduced earlier fire change i.e. fired change on PRE-selected item-press]
and substituted with an alternative check (call to _checkSelectionChange in a hook to 'tab' event)
-- Modified solution I6a8641a6e70e0d970b3db56fabc13f66a01ac400 [that reverts selection on ESC] as it depended on the above change

Change-Id: I53add3b0a209985205c84f3188d52bdd8af57ab6
BCP: 1770589645