[FIX] sap.m.StepInput - wrong calculation when value precision is changed.
After initial rendering of the StepInput typing or programatically setting
of new value with different amount of digits after the dot used to expose
a buggy behavior, where calculation is wrong when +/- button is pressed.

The reason for this was not updating an internal variable holding the
precision of the property <value>.

Small refactoring in _calculateNewValue is done.

(cherry picked from commit c74434f)

Change-Id: I8c493683f16ada0d742d637f92fc4643102dcbdc
BCP: 1780435676
CR-Id: 002075125800004453182017