[FIX] sap.m.Popover: Updated: Show scrollbar when overflowing and out of visible area.
This change enhances the one in 7234648e9805eb51ddfc265aca1b8a3750c58c9b.
The previous one took into account just if the scrolling is enabled,
but not if there's actually some scrolling needed.

With that change the check is if scrolling is enabled AND if the scrolling
actually appears.
Without that change there's a slight offset in the Popover.

BCP: 1780323187

Change-Id: Ie5055fe13d508ca51dff62836981eb44a9b27c99
CR-Id: 002075125800004457782017