[INTERNAL][FIX] DynamicPage test pages adjusted
The DynamicPageUtility.js has gone through automatic
ESlint fixing, which made some of the functions not accessible
from DynamicPage's test pages. This fix wraps them in a global
object (on the window object, to make it global and avoid
another automatic change if declared with var) and adjusts
all pages, which use that utility file.

Change-Id: Ie582c38571bf59965e5e2a5413d2ed5be5d945cb
BCP: 1780438418