[FIX] sap.m.ProgressIndicator: percentValue property correctly validated
Note:
The type of the percentValue property is float.

Issue
String, such as '11' and "110" are treated as not valid
values from the isValidPercentValue function,
which is correct as they are not numbers, but strings.

But, the logic after is confisung.

Having this code block in mind, which is a fallback for invalid values:
if (!isValidPercentValue(fPercentValue)) {
	fNotValidValue = fPercentValue;
	fPercentValue = fPercentValue > 100 ? 100 : 0;
	jQuery.sap.log.warning(this + ": percentValue (" + fNotValidValue + ") is not correct! Setting the percentValue to " + fPercentValue);
}

Case setting percentValue to '110'
- the string is converted to 110, which is bigger than 100 and 100 will be set later.
(fPercentValue = fPercentValue > 100 ? 100 : 0;)

Case setting percentValue to '11'
- the string is converted to 11, which is less than 100 and 0 will be set later.
(fPercentValue = fPercentValue > 100 ? 100 : 0;)

This is confusing because,
using the "110" string seems as it works,
then using "11" is expected to result in 11, not in 0.

Solution 1 (in the change):
I am using the framework validateProperty to delegate the type validation to the framework.
The control now checks only if the value is between 0 and 100  and fallbacks to
a) 100, if the new value is bigger than 100 and
b) 0, if it`s less than 0.

Solution 2 (also possible):
We can change the fallback logic "fPercentValue = fPercentValue > 100 ? 100 : 0;",
which expects to compare only numbers.

Note: validateProperty will throw an error when not valid types of values are used.

Change-Id: Ide7e3ec233862ebae5529d4b624b91823be11ace
BCP: 1780413910