[INTERNAL][FIX] uxap.ObjectPageHeaderContent: ObjectAtrribute appearance fixed
Issue:
In IE only, the ObjectAttribute inside ObjectPageHeaderContent
wraps, although there is enough space to be displayed inone row.

Root cause:
it seems the 'initial' keyword is not recognized by IE.
https://msdn.microsoft.com/en-us/library/hh781508(v=vs.85).aspx#keywords
https://developer.mozilla.org/en-US/docs/Web/CSS/initial

Solution provided:
As the initial value of "display" for "span" (ObjectAttribute is span) is "inline",
I used display: "inline" to resolve the issue.

Change-Id: I7abda39a75d41846e6434b3866f969bfd3c745e4
BCP: 1770487754