[INTERNAL] sap.m.ListBase: Removing tabindex from the root element of the ListBase.
Removing tabindex from the root element of the ListBase and also
cleaning related CSS rules.
Forwarding the focus to the After element when the growing button is
hidden loading all the data.
Due to accessibility issue, the root element is not focusable, hence
changing the same in the NotificationListBase and setting the focut on
the on the list directly.

BCP: 1780447649

Change-Id: I577df6c5b416ebdfdb16e23fb994c8f13019d02f