[INTERNAL][FIX] sap.uxap.ObjectPageSubSection: Ensured renderer sets media class
Background
The ObjectPageSubSection has media-dependent style-class that was set on *after* rendering

Problem:
In the time-slot between (1) flush of the rendered subSection DOM element and (2) call of sunSection.onAfterRendering.
the flushed dom element [of the subSection] does NOT have the media-dependent style-class YET
=> its height is different (as the media-dependent style class affects its height)
=> the scrollHeight of the entire scrollable content of ObjectPageLayout is different (as the subSection is part of the scrollable content)
=> ***the page may [undesirably] ^scroll* as a result*** IF the following happens to be ALSO the case:
  -- the scrollHeight is now smaller [without the media-dependent style class] AND
  -- the current scrollTop [that was set BEFORE the section was rerendered] is bigger than the available scrollHeight
so the page automatically scrolls to a smaller scrollTop.

The wrong scrollPosition will persist if somehing within the rerendered section (inernal grid in this case) makes an operation that causes DOM recalculation
(call to obtain its parent width/height metrics in this case).

Solution:
Fixed by ensuring the media-dependent style-class is set from the renderer
=> so that the flushed dom element has all the classes [that determine its height] already set

Also did a minor refactoring:
changed the place where the listener for media change is attached
(since no need to detach/attach it on before/after rendering,
but sufficient to attach in once on init)

Change-Id: I0ce36fa5315d342846acf20c06e87c1a04ed4bb1
BCP: 1780430231