[FIX]sap.ui.rta: Missing root control should not block application load
If root control could not be retrieved while starting RTA, the
application currently gets stuck in an unresolved promise. This change
rejects the promise and informs the user quickly that RTA could not be
started, also enabling a second attempt that could be successful once
the root control is loaded.

Change-Id: I7312d3c85ee4db020383d4554f8106fa647104d0
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>