[INTERNAL] sap.ui.documentation: Expression binding fix
The problem:
API Reference is unable to show control details and is stuck in loading.

The root cause:
During Maven build expression binding variables (which has the same
syntax) are treated as a Maven properties instead and replaced with the
values provided by Maven which is resulting in a undefined value passed
as a local expression.

The solution:
Use more verbose syntax for expression bindings which are with the same
name as Maven variables.

Change-Id: I21e97fbe41eae370e676503feacb4dc48b39846f