[FEATURE] sap.m.ListBase: new contextMenu aggregation
New aggregation contextMenu and corresponding event
beforeOpenContextMenu is added to the sap.m.ListBase. The event is fired
when the context menu is opened on the list items. Using this
aggregation the application developers can now create a custom context
menu for the controls that extends the sap.m.ListBase control.

Resolves:
JIRA: CPOUIFTEAMB-964

Change-Id: I3ee7c5fb9411c3d35d203ceaa5fb32fc7d470322