[INTERNAL][FIX] NumberFormat: adapt unit test to CLDR version update
With the recent update of the CLDR data, the short-style number format
strings changed for the Spanish locale. 'MRD' is no longer used for
billions and therefore 'M' and 'MRD' are no longer an example for
conflicting units. However, 'mil' (thousands) and 'mil M' (thousands of
millions) can be used as an alternative.

Change-Id: If7698dd13e7a5a044ab040bf41b1e054fb82c3e6