[FIX] DemokitV2: APIRef method parameter properties nesting fixed
The problem:
Method parameter properties are nested on the wrong parameter

The root cause:
Method parameter properties ware nested always after the last parameter
which if we have more parameters with properties in the list they ware
always collected and nested after the last parameter. This is fixed in
api.json transformation with another change but adaptation to custom
control ParamText is needed for the DemoKit to display nesting in a
correct way.

The solution:
Adapt custom ParamText control to handle proper parameter nesting with
new format data from api.json
Keep ParamText work with old format data for events and constructor
usage

Change-Id: I6d4fb39e21ce1d5824a2d2ed2dbfbd18a6a66701
Fixes: https://github.com/SAP/openui5/issues/1798