[INTERNAL] DemokitV2: Fix disappearing methods, events or annotations
The problem:
When navigating between the same kind of entities in the API reference
tree the first event of the symbol viewed sometimes disappear and is
only visible in the summary table. This is not reproducible if you
access the symbol from a deep link. This is also happening for methods
and annotations.

The root cause:
A new lifecycle for summary tables was recently introduced which
introduced the problem. The summary tables are injected to a data bound
list in the ObjectPageSection and when the binding context changes
on navigation the list binding try's to reuse already created items.
The first item to be reused by the first entity is now of a different
type (the summary subSection) and this fails silently skipping the first
'event' and the summary table is recreated and inserted after that.

The solution:
Improving summary sections lifecycle by removing them from their parent
aggregation and destroying them. This is also preventing a memory leak.

Change-Id: I5c0861ee3d202334c3114e9a8df65ae8a3be01a7