[FIX] NumberFormat: percent with custom pattern
Fixed behavior of percent instance in regards to custom pattern.
The format and parse were not symmetric as for percent parsing the
percent pattern was not taken from the provided options but from the
cldr definition. This was done as initially a float instance could also
parse percent values. For the float instance percent parsing is now
tested with additional unit tests.

Change-Id: I7e07a0e5de5ba531eff2d6a07f7398c092d6d720