[FIX] sap.m.RangeSlider: Tooltip exception fixed.
RangeSlider threw exception in an edge case where the slider with tooltip was directly focused on rendering.
This happens when the Slider is the first element on the page on the tab chain and the focus goes directly on it on page load.
The root cause is that RangeSlider is rendered, but stylesheets were not applied.

Note: This issue appears on RangeSlider after the refactoring and it does not impact the older versions.

BCP: 1780507988

Change-Id: Ifb0cfdd12493f3c388e99822bb75a8234973d098