[FEATURE] Control: BusyIndicator animation default
Instead of calculating the BusyIndicator animation
size from DOM (forces a reflow), we added a new
property BusyIndicatorAnimation to handle the 
animation size. 
In most cases the default option should be
sufficient. If the size calculation is still
necessary the property could be set to 'Auto',
or for small sized controls even to 'Small'.
Full screen busy indication should use the 'Large'
option.

Change-Id: I7ffe13d164b70e216584251df01bfa59fd8c362f