[INTERNAL][FIX] sap.m.MultiCombobBox: Fix failing qUnit tests on IE.
BCP: 1780506558

With the fix implemented in the following change Id33e971cf710ba1479d519d575ac0ce7862f13c0, that behavior was changed.
Now when there's an input in the MultiComboBox and it differs of what was
already set, on focus out there's a change event.
As the MultiComboBox clears invalid values on focusout, there's a
change event fired in order to be consistent with "Enter".

Actually the IE behavior was the right one.

Change-Id: Id1947e57cf69c9dcb9f77e2a2de08f5d88fc89b9