[INTERNAL] core.Element: Do not remove DOM synchronously on destroy
 * Reasons:
   - Removing DOM synchronously causes flickering since rendering is
asynchronous. This double reflow-repaint sequence causes performance
issues.
   - Default updateAggregation method destroys all aggregations and
recreate from scratch when grouping and factory functions are used on
the list binding. In the mean time, we are working on new semantic
rendering syntax that can be used for patching DOM nodes. If we remove
DOM nodes synchronously then we cannot patch old control nodes with new
control nodes that are created by the ManagedObject#updateAggregation.

 * Motivations:
   - "KeepDom" parameter of the destroy is used heavily at the
List/Table/ManagedObject/XMLView and we haven't got any issue yet.
   - This change will try to make this default for all controls.

 * Implementation:
   - On destroy we do not remove the control DOM synchronously and we
just let the invalidation happen and wait the next tick of
the RenderManager to remove the DOM of the control.
   - To make this new behavior more compatible we are changing the id of
the control's DOM and all child nodes that starts with the control
ID.(e.g. in XMLView IDs are prefixed)
   - bSuppressInvalidate=true/PopupInterface will remove the DOM
synchronously as before. "KeepDom" is not changed.

 * Side effects:
   - Changing the id of the DOM nodes after destroy helps not to find
control DOM node references with getDomRef() or $() calls but controls
private DOM node references will still be alive.
   - Asynchronous events, like scroll, might still be fired even after
destroy which was not possible before with sync DOM removal therefore
controls should clean their custom events (e.g. registered by jQuery#on)
on exit callback.

Change-Id: Ide04a0e9738e55fa6b13a704c2839d50883718a5